requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
google-generativeai
streamlit
requests
typer
rich
